requirements.txt
aeidon==1.4.1
apturl==0.5.2
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.7.1
blinker==1.4
Brlapi==0.6.7
certifi==2018.8.24
chardet==3.0.4
Click==7.0
colorama==0.3.7
command-not-found==0.3
constantly==15.1.0
cryptography==2.3
cssselect==1.0.3
cupshelpers==1.0
dataclasses==0.6
defer==1.0.6
diff-match-patch==20121119
distro==1.3.0
distro-info===0.21ubuntu2
Django==2.2
django-bootstrap-pagination==1.7.0
django-tinymce==2.8.0
djongo==1.2.32
entrypoints==0.3
html5lib==1.0.1
httplib2==0.11.3
hyperlink==19.0.0
idna==2.6
incremental==17.5.0
iniparse==0.4
keyring==17.1.1
keyrings.alt==3.1.1
l20n==4.0.0a1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.8.0
lxml==4.3.3
macaroonbakery==1.2.1
Mako==1.0.7
MarkupSafe==1.1.0
mongoengine==0.17.0
netifaces==0.10.4
oauthlib==2.1.0
olefile==0.46
parsel==1.5.1
pexpect==4.6.0
phply==1.2.5
Pillow==5.4.1
ply==3.11
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycairo==1.16.2
pycountry==17.5.14
pycrypto==2.6.1
pycups==1.9.73
PyDispatcher==2.0.5
pyenchant==2.0.0
PyGObject==3.32.0
PyHamcrest==1.9.0
PyJWT==1.7.0
pymacaroons==0.13.0
pymongo==3.8.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
PyQt5==5.12.1
pyRFC3339==1.1
python-apt==1.8.4
python-dateutil==2.7.3
python-debian==0.1.34
python-Levenshtein==0.12.0
python-scrapyd-api==2.1.2
pytz==2018.9
pyxdg==0.25
PyYAML==3.13
queuelib==1.5.0
reportlab==3.5.18
requests==2.21.0
requests-unixsocket==0.1.5
Scrapy==1.6.0
scrapy-djangoitem==1.1.1
scrapyd==1.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.8
sqlparse==0.2.4
system-service==0.3
systemd-python==234
translate-toolkit==2.3.1
Twisted==19.2.0
ubuntu-advantage-tools==19.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
vobject==0.9.6.1
w3lib==1.20.0
wadllib==1.3.3
webencodings==0.5.1
xkit==0.0.0
zope.interface==4.6.0
